refactor dauth ... separate authenticator and ratelimiter plugin both have a registry and a null implementation
