+++ authenticator/credentials.go
@@ -1,0 +1,33 @@
+// Copyright 2019 dfuse Platform Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package authenticator
+
+import "go.uber.org/zap"
+
+type Credentials interface {
+	GetLogFields() []zap.Field
+}
+
+type AnonymousCredentials struct {
+	// MUST implement Credentials
+}
+
+func (c *AnonymousCredentials) GetLogFields() []zap.Field {
+	return []zap.Field{
+		zap.String("subject", "anonymous"),
+		zap.String("api_key_id", "anonymous"),
+		zap.String("ip", "0.0.0.0"),
+	}
+}
