+++ authenticator/secret/credentials.go
