+++ go.mod
@@ -1,15 +1,15 @@
 module github.com/dfuse-io/dauth
 
 require (
-	github.com/dfuse-io/derr v0.0.0-20200406214256-c690655246a1
-	github.com/dfuse-io/logging v0.0.0-20200406213449-45fc25dc6a8d
+	github.com/dfuse-io/derr v0.0.0-20200417132224-d333cfd0e9a0
+	github.com/dfuse-io/dfuse-saas-priv v0.0.0-20200527182426-9fa7e3d4cc93
+	github.com/dfuse-io/logging v0.0.0-20200407175011-14021b7a79af
+	github.com/go-redis/redis v6.15.8+incompatible
 	github.com/kr/pretty v0.2.0 // indirect
-	github.com/stretchr/testify v1.4.0
-	go.uber.org/atomic v1.5.1 // indirect
-	go.uber.org/multierr v1.4.0 // indirect
-	go.uber.org/zap v1.14.0
-	google.golang.org/grpc v1.22.1
-	gopkg.in/yaml.v2 v2.2.4 // indirect
+	github.com/stretchr/testify v1.5.1
+	go.uber.org/atomic v1.6.0
+	go.uber.org/zap v1.14.1
+	google.golang.org/grpc v1.26.0
 )
 
 go 1.13
